dashboard: simplify estimations
The naive average is only a problem if you have many measurements close
together. We can alleviate that by naively grouping measurements for a
single day together.

Linear interpolation is not good enough of an approximation in this
case, fuzzier metrics appear to give a better sense of the current
situation and slope.