Simplify schema, update duckdb
Schema simplification. Replace the timestamp (ts) column in weight with
a date column. It is easier to input, easier to deal with, smaller, and
the extra granularity of different times of the day is just noise, I
haven't gotten any insights from that.

Update duckdb. Since the duckdb on disk format is not stable, store the
database as parquet. Import and export the database for every command.
This is snappy enough.